requirements.txt
# Core Framework
fastapi==0.128.0
pydantic==2.12.5
uvicorn[standard]==0.38.0
starlette==0.50.0

# Database & Cache
redis==7.1.0
sqlalchemy==2.0.45
alembic==1.17.2
asyncpg==0.31.0
psycopg2-binary==2.9.11

# Third-party Integrations
lark-oapi==1.4.24
requests==2.32.5
httpx==0.28.1

# Utilities
psutil==7.2.1
python-dotenv==1.2.1
python-json-logger==4.0.0
prometheus-client==0.23.1

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1

# Platform Specific
setproctitle==1.3.3; sys_platform == 'linux'
